pyproject.toml
[tool.poetry]
name = "music-store"
version = "0.1.0"
description = ""
authors = ["Almaz"]
readme = "README.md"
packages = [{include = "music_store"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.15"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.10.8"}
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
isort = "^5.12.0"
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
